Additional fixes for QMP & BG/L
git-svn-id: file:///home/chulwoo/CPS/repo/CVS2SVN/cps/trunk@2585 fb734a29-28db-4bc5-88a3-3cb5ee595545

+++ cps_pp/src/comms/bgl/cbuf/cram.C
@@ -4,12 +4,12 @@ CPS_START_NAMESPACE
 //  CVS keywords
 //
 //  $Author: chulwoo $
-//  $Date: 2006-12-14 17:53:43 $
-//  $Header: /home/chulwoo/CPS/repo/CVS/cps_only/cps_pp/src/comms/bgl/cbuf/cram.C,v 1.2 2006-12-14 17:53:43 chulwoo Exp $
-//  $Id: cram.C,v 1.2 2006-12-14 17:53:43 chulwoo Exp $
+//  $Date: 2007-01-11 22:48:12 $
+//  $Header: /home/chulwoo/CPS/repo/CVS/cps_only/cps_pp/src/comms/bgl/cbuf/cram.C,v 1.3 2007-01-11 22:48:12 chulwoo Exp $
+//  $Id: cram.C,v 1.3 2007-01-11 22:48:12 chulwoo Exp $
 //  $Name: not supported by cvs2svn $
 //  $Locker:  $
-//  $Revision: 1.2 $
+//  $Revision: 1.3 $
 //  $Source: /home/chulwoo/CPS/repo/CVS/cps_only/cps_pp/src/comms/bgl/cbuf/cram.C,v $
 //  $State: Exp $
 //
@@ -21,6 +21,6 @@ CPS_START_NAMESPACE
 //Workstation version only
 #ifndef _TARTAN
 int CRAM_SCRATCH_INTS[CRAM_SCRATCH_SIZE] ;
-unsigned int CRAM_SCRATCH_ADDR = (unsigned int)CRAM_SCRATCH_INTS ;
+unsigned long CRAM_SCRATCH_ADDR = (unsigned long)CRAM_SCRATCH_INTS ;
 #endif
 CPS_END_NAMESPACE
